metadata/scripts/requirements.txt
GitPython>=3.1
